+++ packages/squiggle-lang/scripts/bench-map-reduce.mjs
@@ -1,15 +1,6 @@
 #!/usr/bin/env node
 import { SqProject } from "@quri/squiggle-lang";
-
-const measure = (cb, times = 1) => {
-  const t1 = new Date();
-
-  for (let i = 1; i <= times; i++) {
-    cb();
-  }
-  const t2 = new Date();
-  return (t2 - t1) / 1000;
-};
+import { measure } from "./lib.mjs";
 
 const maxP = 5;
