+++ packages/squiggle-lang/scripts/run.mjs
@@ -1,18 +1,10 @@
 #!/usr/bin/env node
-import { SqProject } from "@quri/squiggle-lang";
-
-const project = SqProject.create();
+import { run } from "./lib.mjs";
 
 const src = process.argv[2];
 if (!src) {
   throw new Error("Expected src");
 }
 console.log(`Running ${src}`);
-project.setSource("a", src);
-project.run("a");
 
-const result = project.getResult("a");
-console.log(result.tag, result.value.toString());
-
-const bindings = project.getBindings("a");
-console.log(bindings.asValue().toString());
+run(src);
